Move QRcode replacement magic into separate file
Make the QRcode replacement magic an event handler for the Reveal
ready event, and rather than ID reference, have it honor the
"qrcode" class.

Thus, you're able to do (in Markdown):

<https://myterrificsite.ex4mpl3.com/foo> <!-- .element class="qrcode" -->

... and that link gets replaced by a 400x400 QR code.

One thing that's a bit ugly at this stage is that if you put a link on
the very first slide, then the QR code falls off the screen edge when
the slides are first rendered. Changing to a different slide, or even
just resizing the browser window, will center the QR code though.