Flip the preference for the notes separator
When viewing a Markdown file in GitHub or GitLab, HTML comments
are treated as such. That means when using "<!-- Note -->" as the
notes separator, this is removed from the output, so that someone
who reads the Markdown online has a cleaner reading flow.

Thus, flip the default for the notes separator from upstream's
"Note:" to the HTML comment.