Expose the menu plugin's theme switch
The black-on-white theme is not for everyone: someone dealing with
migraine may be in considerable discomfort from looking at a large
expanse of white, for example.

Expose the menu plugin's theme switcher, so that it becomes possible
to select the theme from the pop-out menu (activated by clicking
the hamburger menu, or typing the "m" key).